+++ game-disks/ur-dead.js
@@ -837,134 +837,7 @@ There's a bearded skeleton by the sign. He seems to want to TALK.`,
     },
   ],
   methods: {
-    utils: {
-      // saves text from memory to disk
-      saveFile: (content, filename) => {
-        const a = document.createElement('a');
-        const file = new Blob([content], {type: 'text/plain'});
-
-        a.href = URL.createObjectURL(file);
-        a.download = filename;
-        a.click();
-
-        URL.revokeObjectURL(a.href);
-      },
-      // creates input element to open file prompt (allows user to load exported game from disk)
-      openFile: () => {
-        const input = document.createElement('input');
-        input.setAttribute('type', 'file');
-        input.click();
-
-        return input;
-      },
-      // asserts the command is not save, load, import or export, nor blank (could use a better name...)
-      isNotMeta: (cmd) => !cmd.toLowerCase().startsWith('save')
-        && !cmd.toLowerCase().startsWith('load')
-        && !cmd.toLowerCase().startsWith('export')
-        && !cmd.toLowerCase().startsWith('import')
-        && cmd !== '',
-      // applies string representing an array of input strings (used for loading saved games)
-      applyInputs(string) {
-        let ins = [];
-
-        // parse, filtering out the save/load commands & empty strings
-        try {
-          ins = JSON.parse(string).filter(disk.methods.utils.isNotMeta);
-        } catch(err) {
-          println(`An error occurred. See error console for more details.`);
-          console.error(`An error occurred while attempting to parse text-engine inputs.
-            Inputs: ${string}
-            Error: ${err}`);
-          return;
-        }
-
-        while (ins.length) {
-          applyInput(ins.shift());
-        }
-      },
-    },
     commands: {
-      // override help to include import/export
-      help() {
-        const instructions = `The following commands are available:
-          LOOK:           'look at key'
-          TAKE:           'take book'
-          GO:             'go north'
-          USE:            'use door'
-          TALK:           'talk to mary'
-          ITEMS:          list items in the room
-          INV:            list inventory items
-          SAVE/LOAD:      save current game, or load a saved game (in memory)
-          IMPORT/EXPORT:  save current game, or load a saved game (on disk)
-          HELP:   this help menu
-        `;
-        println(instructions);
-      },
-      // overridden save command stores player input history
-      // (optionally accepts a name for the save)
-      save: (name = 'save') => {
-        localStorage.setItem(name, JSON.stringify(inputs));
-        const line = name.length ? `Game saved as "${name}".` : `Game saved.`;
-        println(line);
-      },
-      // overridden load command reapplies inputs from saved game
-      // (optionally accepts a name for the save)
-      load: (name = 'save') => {
-        if (inputs.filter(disk.methods.utils.isNotMeta).length > 2) {
-          println(`At present, you cannot load in the middle of the game. Please reload the browser, then run the **LOAD** command again.`);
-          return;
-        }
-
-        let save = localStorage.getItem(name);
-
-        if (!save) {
-          println(`Save file not found.`);
-          return;
-        }
-
-        disk.methods.utils.applyInputs(save);
-
-        const line = name.length ? `Game "${name}" was loaded.` : `Game loaded.`;
-        println(line);
-      },
-      // export current game to disk (optionally accepts a filename)
-      export(name) {
-        const filename = `${name.length ? name : 'urdead'}.txt`;
-        disk.methods.utils.saveFile(JSON.stringify(inputs), filename);
-        println(`Game exported to "${filename}".`);
-      },
-      // import a previously exported game from disk
-      import() {
-        if (inputs.filter(disk.methods.utils.isNotMeta).length > 2) {
-          println(`At present, you cannot load in the middle of the game. Please reload the browser, then run the **IMPORT** command again.`);
-          return;
-        }
-
-        const input = disk.methods.utils.openFile();
-        input.onchange = () => {
-          const fr = new FileReader();
-          const file = input.files[0];
-
-          // register file loaded callback
-          fr.onload = () => {
-            // load the game
-            disk.methods.utils.applyInputs(fr.result);
-            println(`Game "${file.name}" was loaded.`);
-            input.remove();
-          };
-
-          // register error handling
-          fr.onerror = (event) => {
-            println(`An error occured loading ${file.name}. See console for more information.`);
-            console.error(`Reader error: ${fr.error}
-              Reader error event: ${event}`);
-            input.remove();
-          };
-
-          // attempt to load the text from the selected file
-          fr.readAsText(file);
-        };
-      },
       play: () => println(`You're already playing a game.`),
       // set player's name
       name: (arg) => {
